fix(server): properly assign user data on update
Prior to this change, there was an issue where updating a user's
information was not being persisted correctly.

This change properly assigns the updated `age` to the user record `age`
field, not the `id`.  Assigning `age` to `id` would result in the target
record not being updated, and instead a new record with the incorrect
`id` being created.

Fixes #1