feat: add notes to users
Prior to this change, there were no notes on users.

This change adds the database, backend, and frontend code to enable
adding notes to users. The notes will be displayed in the user list page
in reverse chronological order, with their timestamps. A "Add Note"
button is added in the user list for each user, and when clicked
presents an inline form to add the note. The existing behavior
of requiring a refresh to see the updated record is still present. Some
styling has been applied to make the notes visually distinct.

Completes #2