feat: add user record page
Prior to this change, there was no page that showed just a single user
record.

This change adds a user record page, accessible from `/user/<userId>`.
This includes a new `get` endpoint in the backend, and adds simple
react-router page routing in the frontend to enable the custom URL for
the record page. We also factor out a shared `UserNotes` component for
displaying the notes list in both the user list and user profile page.

Completes #3